Seperate App and Content Component
